backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
transformers==4.44.0
torch
torchvision
torchaudio
opencv-python==4.10.0.84
Pillow==10.4.0
ezdxf==1.3.0
numpy==1.26.4
pydantic==2.8.2
python-dotenv==1.0.1
sqlalchemy==2.0.30
pymysql==1.1.1
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
passlib==1.7.4